Echo_Backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
pymongo
python-dotenv
markdown
weasyprint
python-jose[cryptography]
passlib[bcrypt]
